Reload page if game fails to load
Sometimes after an automatic refresh, the servers error out on providing a page or script and the game hangs until the next automatic refresh. This causes the script to reload the page again if the game is not functional within 30 seconds.

This may need further adjustments for morning lag compensation when the servers reset as that will probably have everyone's game lagging in unexpected ways.